new version with movable y-axis
scoreGT problem solved!!!
The mistake I made is that I forgot to delete the basket script for Basket prefab, which result in scoreGT as null when collision because it didn't have the script to initialize scoreGT. I separate the basket script into two. That Basket.cs manage the parent empty game object basket that handles movement, while Basketpref.cs handle collision and it applies to basket prefab. The problem solved!